+++ KursovayaEngine2/Tools/BinarySearch.h
@@ -6,15 +6,15 @@
 ///compareFunc should check if for example p1<p2, p1 and p2 are in order of array
 ///equalFunc should check if params are equal
 template<typename ArrType, typename ValType>
-unsigned int BinarySearch(const ArrType* arr, unsigned int arrLen, ValType&& val, 
+size_t BinarySearch(const ArrType* arr, size_t arrLen, ValType&& val, 
 	bool(*compareFunc)(const std::remove_reference_t<ValType>&, const ArrType&),
 	bool(*equalFunc)(const std::remove_reference_t<ValType>&, const ArrType&)) {
 
 	if (arrLen == 0) return 1;
 	
-	unsigned int start = 0; unsigned int end = arrLen - 1;
+	size_t start = 0; size_t end = arrLen - 1;
 	while (true) {
-		unsigned int ind = start + (end - start) / 2;
+		size_t ind = start + (end - start) / 2;
 		if (compareFunc(val, arr[ind])) {
 			if (ind == 0) return 0;
 			else if (not compareFunc(val, arr[ind - 1])) return equalFunc(val, arr[ind - 1]) ? (ind - 1) : ind;
@@ -40,14 +40,14 @@ unsigned int BinarySearch(const ArrType* arr, unsigned int arrLen, ValType&& val
 ///if function returns arrLen+1 then it failed for some reason
 ///compareFunc should check if for example p1<p2, p1 and p2 are in order of array
 template<typename ArrType, typename ValType>
-unsigned int BinarySearch(const ArrType* arr, unsigned int arrLen, ValType&& val, 
+size_t BinarySearch(const ArrType* arr, size_t arrLen, ValType&& val, 
 	bool(*compareFunc)(const std::remove_reference_t<ValType>&, const ArrType&)) {
 
 	if (arrLen == 0) return 1;
 	
-	unsigned int start = 0; unsigned int end = arrLen - 1;
+	size_t start = 0; size_t end = arrLen - 1;
 	while (true) {
-		unsigned int ind = start + (end - start) / 2;
+		size_t ind = start + (end - start) / 2;
 		if (compareFunc(val, arr[ind])) {
 			if (ind == 0) return 0;
 			else if (not compareFunc(val, arr[ind - 1])) return ind;
